// src/types/enums.ts
export enum Languages {
  c = 'C',
  cpp = 'Cpp',
  cuda = 'Cuda',
}

export enum CompilerSystems {
  gcc = 'gcc',
  clang = 'clang',
  msvc = 'msvc',
}

export enum Debuggers {
  lldb = 'lldb',
  gdb = 'gdb',
}

export enum OperatingSystems {
  windows = 'windows',
  linux = 'linux',
  mac = 'macos',
}

export enum Architectures {
  x86 = 'x86',
  x64 = 'x64',
  ARM64 = 'ARM64',
}

export enum Builds {
  debug = 'Debug',
  release = 'Release',
}